Inheritance classes and file manager both have now their own file. reduced length of module init
